# apps/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lockdev SaaS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/organizations/{org_id}": {
      "get": {
        "summary": "Get Organization",
        "description": "Get organization details. Requires membership.",
        "operationId": "get_organization_api_v1_organizations__org_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Organization",
        "description": "Update organization. Requires ADMIN role.",
        "operationId": "update_organization_api_v1_organizations__org_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "Real-Time Events"
        ],
        "summary": "Get Events",
        "description": "Server-Sent Events stream for real-time updates.\n\nThis endpoint establishes a long-lived HTTP connection for\nserver-to-client push notifications.\n\nThe client will receive events such as:\n- notification.new: New notification created\n- message.new: New message in subscribed thread\n- appointment.reminder: Upcoming appointment reminder\n- appointment.updated: Appointment rescheduled/cancelled\n- document.processed: Document finished OCR/virus scan\n- document.scan_complete: Virus scan completed\n- consent.required: New consent document available\n- call.incoming: Incoming call (Call Center)\n- call.completed: Call ended\n- heartbeat: Keep-alive (every 30 seconds)\n\nAuthentication can be provided via:\n- Authorization header: Bearer {token}\n- Query parameter: ?token={token}\n\nReturns:\n    StreamingResponse with text/event-stream content type",
        "operationId": "get_events_api_v1_events_get",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT token for authentication",
              "title": "Token"
            },
            "description": "JWT token for authentication"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/organizations/{org_id}": {
      "get": {
        "tags": [
          "Real-Time Events"
        ],
        "summary": "Get Org Events",
        "description": "Organization-scoped SSE stream for admin views.\n\nThis endpoint provides organization-level events in addition\nto standard user events. Requires ADMIN or SUPER_ADMIN role.\n\nAdditional event types:\n- member.joined: New member accepted invitation\n- member.removed: Member removed from organization\n- subscription.updated: Subscription status changed\n- license.expiring: Provider license expiring soon\n\nArgs:\n    org_id: Organization ID\n\nReturns:\n    StreamingResponse with text/event-stream content type\n\nRaises:\n    HTTPException: If user is not admin of the organization",
        "operationId": "get_org_events_api_v1_events_organizations__org_id__get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JWT token for authentication",
              "title": "Token"
            },
            "description": "JWT token for authentication"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrganizationRead": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name"
          },
          "tax_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Id"
          },
          "settings_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Settings Json"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "stripe_customer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Customer Id"
          },
          "subscription_status": {
            "type": "string",
            "title": "Subscription Status"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "member_count": {
            "type": "integer",
            "title": "Member Count"
          },
          "patient_count": {
            "type": "integer",
            "title": "Patient Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "stripe_customer_id",
          "subscription_status",
          "is_active",
          "member_count",
          "patient_count",
          "created_at",
          "updated_at"
        ],
        "title": "OrganizationRead"
      },
      "OrganizationUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "tax_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Tax Id"
          },
          "settings_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settings Json"
          }
        },
        "type": "object",
        "title": "OrganizationUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}